Move ProviderInformation to f27x
